Move fusion to core
